fix: rename to no-unbound-methods
